6.1.6 Создаст еще один абстрактный класс Smart в который вынесет логику обновления локальных данных и перерисовки для интрактивных компонентов. Унаследует Popup от Smart, а Smart от Abstract. Вызов метода restoreHandlers переместит в метод updateElement для удобства.
